[chore] [pdata] Move constants to separate files
