+++ pdata/pmetric/aggregation_temporality_test.go
@@ -1,0 +1,28 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package pmetric // import "go.opentelemetry.io/collector/pdata/pmetric"
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestAggregationTemporalityString(t *testing.T) {
+	assert.Equal(t, "Unspecified", AggregationTemporalityUnspecified.String())
+	assert.Equal(t, "Delta", AggregationTemporalityDelta.String())
+	assert.Equal(t, "Cumulative", AggregationTemporalityCumulative.String())
+	assert.Equal(t, "", (AggregationTemporalityCumulative + 1).String())
+}
