+++ pdata/ptrace/traces.go
@@ -17,7 +17,6 @@ package ptrace // import "go.opentelemetry.io/collector/pdata/ptrace"
 import (
 	"go.opentelemetry.io/collector/pdata/internal"
 	otlpcollectortrace "go.opentelemetry.io/collector/pdata/internal/data/protogen/collector/trace/v1"
-	otlptrace "go.opentelemetry.io/collector/pdata/internal/data/protogen/trace/v1"
 )
 
 // Traces is the top-level struct that is propagated through the traces pipeline.
@@ -66,73 +65,4 @@ func (ms Traces) SpanCount() int {
 // ResourceSpans returns the ResourceSpansSlice associated with this Metrics.
 func (ms Traces) ResourceSpans() ResourceSpansSlice {
 	return newResourceSpansSlice(&ms.getOrig().ResourceSpans)
-}
-
-// SpanKind is the type of span. Can be used to specify additional relationships between spans
-// in addition to a parent/child relationship.
-type SpanKind int32
-
-const (
-	// SpanKindUnspecified represents that the SpanKind is unspecified, it MUST NOT be used.
-	SpanKindUnspecified = SpanKind(otlptrace.Span_SPAN_KIND_UNSPECIFIED)
-	// SpanKindInternal indicates that the span represents an internal operation within an application,
-	// as opposed to an operation happening at the boundaries. Default value.
-	SpanKindInternal = SpanKind(otlptrace.Span_SPAN_KIND_INTERNAL)
-	// SpanKindServer indicates that the span covers server-side handling of an RPC or other
-	// remote network request.
-	SpanKindServer = SpanKind(otlptrace.Span_SPAN_KIND_SERVER)
-	// SpanKindClient indicates that the span describes a request to some remote service.
-	SpanKindClient = SpanKind(otlptrace.Span_SPAN_KIND_CLIENT)
-	// SpanKindProducer indicates that the span describes a producer sending a message to a broker.
-	// Unlike CLIENT and SERVER, there is often no direct critical path latency relationship
-	// between producer and consumer spans.
-	// A PRODUCER span ends when the message was accepted by the broker while the logical processing of
-	// the message might span a much longer time.
-	SpanKindProducer = SpanKind(otlptrace.Span_SPAN_KIND_PRODUCER)
-	// SpanKindConsumer indicates that the span describes consumer receiving a message from a broker.
-	// Like the PRODUCER kind, there is often no direct critical path latency relationship between
-	// producer and consumer spans.
-	SpanKindConsumer = SpanKind(otlptrace.Span_SPAN_KIND_CONSUMER)
-)
-
-// String returns the string representation of the SpanKind.
-func (sk SpanKind) String() string {
-	switch sk {
-	case SpanKindUnspecified:
-		return "Unspecified"
-	case SpanKindInternal:
-		return "Internal"
-	case SpanKindServer:
-		return "Server"
-	case SpanKindClient:
-		return "Client"
-	case SpanKindProducer:
-		return "Producer"
-	case SpanKindConsumer:
-		return "Consumer"
-	}
-	return ""
-}
-
-// StatusCode mirrors the codes defined at
-// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/api.md#set-status
-type StatusCode int32
-
-const (
-	StatusCodeUnset = StatusCode(otlptrace.Status_STATUS_CODE_UNSET)
-	StatusCodeOk    = StatusCode(otlptrace.Status_STATUS_CODE_OK)
-	StatusCodeError = StatusCode(otlptrace.Status_STATUS_CODE_ERROR)
-)
-
-// String returns the string representation of the StatusCode.
-func (sc StatusCode) String() string {
-	switch sc {
-	case StatusCodeUnset:
-		return "Unset"
-	case StatusCodeOk:
-		return "Ok"
-	case StatusCodeError:
-		return "Error"
-	}
-	return ""
 }
